+++ docs/conf.py
@@ -20,7 +20,7 @@
 project = 'freeproxy'
 copyright = '2022-2030, Zhenchao Jin'
 author = 'Zhenchao Jin'
-release = '0.1.5'
+release = '0.1.6'
 
 # -- General configuration ---------------------------------------------------