+++ docs/Changelog.md
@@ -1,5 +1,6 @@
 # Release Log
 
+- 2025-10-18: Released pyfreeproxy v0.1.6 — new sources `QiyunipProxiedSession`, `ProxyhubProxiedSession` and `ProxydbProxiedSession` added.
 - 2025-10-17: Released pyfreeproxy v0.1.5 — Code cleanup, deprecated/invalid proxy sources removed, new sources added.
 - 2022-03-25: Released pyfreeproxy v0.1.4 — Switched license; added author information printout.
 - 2022-02-21: Released pyfreeproxy v0.1.1 to 0.1.3 — Added two APIs, `getrandomproxysession` and `getrandomproxy`; improved robustness.